+++ R/zzz.R
@@ -3,9 +3,6 @@
 .onAttach <- function(libname, pkgname) {
   bundle <- installed_bundle_manifest()
   build <- installed_build_manifest()
-  if (!build_is_current(bundle, build)) {
-    build <- installed_build_manifest(.libPaths()[[1L]])
-  }
   if (build_is_current(bundle, build)) {
     packageStartupMessage(
       cli::format_inline(